+++ src/main/java/uk/gov/hmcts/reform/notificationservice/service/NotificationMessageParser.java
@@ -7,7 +7,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 import uk.gov.hmcts.reform.notificationservice.exception.InvalidMessageException;
-import uk.gov.hmcts.reform.notificationservice.model.request.incomming.NotificationMsg;
+import uk.gov.hmcts.reform.notificationservice.model.in.NotificationMsg;
 
 import java.io.IOException;
 import java.util.List;
